+++ api/help-walkthrough-media.ts
@@ -1,0 +1,3 @@
+import { createHelpWalkthroughMediaHandler } from '../server/helpWalkthroughMedia.js';
+
+export default { fetch: createHelpWalkthroughMediaHandler() };